move postgres config file to root foler and rename as posgres_config.js
